LPD-86046 Move JWTTokenUtil to impl and replace API usage with interface
